feat: refactor logic to use default in the before each boot
